+++ extensions/obsidian-bookmarks/src/hooks/use-link-form.tsx
@@ -7,6 +7,7 @@ export interface LinkFormState {
   values: {
     title: string;
     url: string;
+    favicon: string;
     tags: string[];
     description: string;
   };
@@ -68,6 +69,7 @@ export default function useLinkForm(initialValues: Partial<LinkFormState["values
     hasUpdatedWithLink: false,
     values: {
       description: "",
+      favicon: "",
       tags: [],
       title: "",
       url: "",