+++ extensions/obsidian-bookmarks/src/types/preferences.tsx
@@ -23,4 +23,5 @@ export interface Preferences {
   useBrowserExtension: boolean;
   searchRecursively: boolean;
   checkDuplicates: boolean;
+  faviconField?: string;
 }